# accounts/abi: improve unpack performance