--- a/accounts/abi/argument.go
+++ b/accounts/abi/argument.go
@@ -182,10 +182,20 @@ func (arguments Arguments) copyTuple(v interface{}, marshalledValues []interface
 // without supplying a struct to unpack into. Instead, this method returns a list containing the
 // values. An atomic argument will be a list with one element.
 func (arguments Arguments) UnpackValues(data []byte) ([]interface{}, error) {
-	nonIndexedArgs := arguments.NonIndexed()
-	retval := make([]interface{}, 0, len(nonIndexedArgs))
+	size := 0
+	for _, arg := range arguments {
+		if arg.Indexed {
+			continue
+		}
+		size++
+	}
+
+	retval := make([]interface{}, 0, size)
 	virtualArgs := 0
-	for index, arg := range nonIndexedArgs {
+	for index, arg := range arguments {
+		if arg.Indexed {
+			continue
+		}
 		marshalledValue, err := toGoType((index+virtualArgs)*32, arg.Type, data)
 		if err != nil {
 			return nil, err